+++ src/form/b-form/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-06-??)
+
+#### :bug: Bug Fix
+
+* Fixed caching of form values
+
 ## v3.0.0-rc.199 (2021-06-16)
 
 #### :boom: Breaking Change